// frontend/src/schemas/auth/register-schema.ts
import { z } from 'zod'

export const registerSchema = z.object({
  name: z.string().min(1, 'Nome é obrigatório'),
  email: z.string().email('Email inválido'),
  password: z.string().min(6, 'Senha deve ter pelo menos 6 caracteres'),
})

export type RegisterSchema = z.infer<typeof registerSchema>